requirements.txt
langchain
openai
streamlit
scikit-learn
PyPDF2
python-docx
beautifulsoup4
python-dotenv
langchain_community
spacy
pandas
numpy
xgboost
joblib
fastapi
uvicorn
python-multipart
pydantic
groq
#python -m spacy download en_core_web_sm
#pip install nltk
# numpy==1.26.4
#pandas==2.2.2
#scikit-learn==1.4.2
#joblib==1.4.2
#nltk==3.8.1
#spacy==3.7.5
#fastapi==0.111.0
#uvicorn==0.30.1
#python-dotenv==1.0.1
#xgboost==2.0.3
#pydantic==2.7.4
